Fix statistics overflow: uint32_t counts and incremental average
measurementCount/errorCount were uint16_t — wrapped at 65535 (~4min at
250fps). avgDistance used a multiply-then-divide formula that overflows
uint32_t on long-running sensors. Switch to incremental average.

+++ DTS6012M_UART.h
@@ -173,8 +173,8 @@ struct DTSStatistics {
   uint16_t minDistance;
   uint16_t maxDistance;
   uint32_t avgDistance;
-  uint16_t measurementCount;
-  uint16_t errorCount;
+  uint32_t measurementCount;
+  uint32_t errorCount;
 };
 
 class DTS6012M_UART {
